lint: 放开第一批 Ruff 规则
启用空白、非法转义、裸 except、类型比较、未定义名称、未使用变量和少量 RUF 规则。

清理对应历史问题，包括字符串尾随空白、正则转义、OCR RPC 裸异常、NumPy 布尔数组比较、未使用局部变量和部分显式转换写法。

保持全仓库 ruff、ty、compileall 和 diff check 通过，为后续继续收紧 Ruff 基线打底。

+++ module/map/camera.py
@@ -40,7 +40,7 @@ class Camera(MapOperation):
             bool: if camera moved.
         """
         vector = np.array(vector)
-        name = "MAP_SWIPE_" + "_".join([str(int(round(x))) for x in vector])
+        name = "MAP_SWIPE_" + "_".join([str(round(x)) for x in vector])
         if np.any(np.abs(vector) > self.config.MAP_SWIPE_DROP):
             # 地图格子滑动倍率按 minitouch 固定。
             distance = self.view.swipe_base * self.config.MAP_SWIPE_MULTIPLY_MINITOUCH
